# public/openapi.yaml
openapi: 3.0.3
info:
  title: Property CRM Demo API
  version: 1.0.0
  description: |
    Demo REST API for the Property Management Front plugin.
    Send header `Authorization: Bearer <token>` (see Back Office View → API docs tab at `/crm?tab=api`; env `NEXT_PUBLIC_DEMO_API_TOKEN`).

    **Single records (per contact email):** For each nested object type (properties, quotes, cases, inquiries, contracts, attachments, work orders, invoices, pets, policies, claims),
    use **GET** `.../{id}` to read one record and **POST** `.../{id}` (or **PUT**) with a JSON body to create or replace that id.
    Collection **POST** `.../cases` (etc.) also upserts by `id` in the body. **Policyholder** and **cover** are singletons: **GET**/**PUT** on `.../policyholder` and `.../cover` only. Timeline events use **index** instead of id (see `/timeline/{index}`).
servers:
  - url: https://front-plugin-demo-app.vercel.app
    description: Production (Vercel)
  - url: http://localhost:3000
    description: Local Next.js dev
paths:
...
  /api/contacts/{email}/work-orders:
    post:
      summary: Create or upsert work order (by body id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrder"
      responses:
        "201":
          description: Updated contact with work orders
        "404":
          description: Contact not found
  /api/contacts/{email}/work-orders/{id}:
    get:
      summary: Get one work order by id
      responses:
        "200":
          description: Work order
        "404":
          description: Contact or work order not found
    post:
      summary: Create or upsert work order (path id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrder"
      responses:
        "201":
          description: Updated contact with work orders
        "404":
          description: Contact not found
    put:
      summary: Upsert work order with path id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrder"
      responses:
        "200":
          description: Updated contact
    delete:
      summary: Remove work order
      responses:
        "200":
          description: Updated contact
  /api/contacts/{email}/invoices:
    post:
      summary: Create or upsert invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Invoice"
      responses:
        "201":
          description: Updated contact
  /api/contacts/{email}/invoices/{id}:
    parameters:
      - name: email
        in: path
        required: true
        schema: { type: string }
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get one invoice by id
      responses:
        "200":
          description: Invoice JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Contact or invoice not found
    post:
      summary: Create or upsert invoice; path id merged into body (201)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Invoice"
      responses:
        "201":
          description: Updated contact
        "400":
          description: Validation error
        "404":
          description: Contact not found
    put:
      summary: Upsert invoice (200)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Invoice"
      responses:
        "200":
          description: Updated contact
        "400":
          description: Validation error
        "404":
          description: Contact not found
    delete:
      summary: Remove invoice
      responses:
        "200":
          description: Updated contact
        "404":
          description: Contact not found
...
  /api/contacts/{email}/cover:
    get:
      summary: Get cover (limits, excess, exclusions)
      responses:
        "200":
          description: Cover JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cover"
        "404":
          description: Contact not found
    put:
      summary: Replace cover; returns full contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cover"
      responses:
        "200": { description: Updated contact }
        "400": { description: Validation error }
        "404": { description: Contact not found }
...
components:
  schemas:
...
    CoverExcess:
      type: object
      required: [fixed, coInsurance]
      properties:
        fixed: { type: number }
        coInsurance: { type: string }
    Cover:
      type: object
      required:
        - vetFeeLimit
        - vetFeeLimitType
        - remainingLimitThisYear
        - excess
        - complementaryTreatment
        - dental
        - thirdPartyLiability
        - exclusions
      properties:
        vetFeeLimit: { type: number }
        vetFeeLimitType: { type: string }
        remainingLimitThisYear: { type: number }
        excess: { $ref: "#/components/schemas/CoverExcess" }
        complementaryTreatment: { type: number }
        dental: { type: number }
        thirdPartyLiability: { type: number }
        exclusions:
          type: array
          items: { type: string }
...
    WorkOrder:
      type: object
      required: [id, title, type, status]
      properties:
        id: { type: string }
        title: { type: string }
        type:
          type: string
          enum: [maintenance, repair, inspection, cleaning]
        status:
          type: string
          enum: [scheduled, in_progress, completed]
        scheduledFor: { type: string }
        propertyId: { type: string }
    Invoice:
      type: object
      required: [id, reference, amount, currency, status]
      properties:
        id: { type: string }
        reference: { type: string }
        amount: { type: number }
        currency: { type: string }
        status:
          type: string
          enum: [draft, pending, paid, overdue]
        dueDate: { type: string }
...